More separation of declaration and definition
