+++ firmware/AVR-Source/Pyr0_Piezo_Sensor_v2.x.x/src/pP_config.h
@@ -5,15 +5,15 @@
 
 #define GAIN_FACTOR_DEFAULT 2 // Gain adjustment factor. 0=3x, 1=3.5x, 2=4.33x, 3=6x, 4=11x
 #define GAIN_FACTOR_ADDRESS 0
-int GAIN_FACTOR = GAIN_FACTOR_DEFAULT;
+extern int GAIN_FACTOR;
 
 #define FOLLOWER_THRESHOLD_DEFAULT 1450 // Voltage follower default voltage in mV
 #define FOLLOWER_THRESHOLD_ADDRESS 4
-int followerThrs = 1450;
+extern int followerThrs;
 
 #define COMP_THRESHOLD_DEFAULT 2850 // Comparatore Vref default voltage in mV
 #define COMP_THRESHOLD_ADDRESS 8
-int compThrs = COMP_THRESHOLD_DEFAULT;
+extern int compThrs;
 
 #ifndef InitCount
 #define InitCount 6 // Number of times to blink the LED on start
@@ -21,29 +21,29 @@ int compThrs = COMP_THRESHOLD_DEFAULT;
 
 #define LOOP_DUR_DEFAULT 50 // duration of time between ADC checks and other loop functions
 #define LOOP_DUR_ADDRESS 12
-int LOOP_DUR = LOOP_DUR_DEFAULT;
+extern int LOOP_DUR;
 
 #define TRG_DUR_DEFAULT 20 // duration of the Z-axis pulse sent, in ms
 #define TRG_DUR_ADDRESS 16
-int TRG_DUR = TRG_DUR_DEFAULT;
+extern int TRG_DUR;
 
 #define HYST_DEFAULT 20
 #define HYST_ADDRESS 20
-int Hyst = HYST_DEFAULT; // Hysteresis value for ADC measurements
+extern int Hyst; // Hysteresis value for ADC measurements
 
 #define LOGIC_DEFAULT 1
 #define LOGIC_ADDRESS 32
-int LOGIC = LOGIC_DEFAULT; // Trigger logic scheme, Active LOW is default
+extern int LOGIC; // Trigger logic scheme, Active LOW is default
 
 #define PZDET_DEFAULT 0
 #define PZDET_ADDRESS 26
-int PZDET = PZDET_DEFAULT; // Enable or disable piezo connection detection, default is off
+extern int PZDET; // Enable or disable piezo connection detection, default is off
 
-int Debug = 0;
+extern int Debug;
 
 #define VM_CONST_ADDRESS 28
 #define VM_CONST_DEFAULT 1125300L
-long voltMeterConstant = VM_CONST_DEFAULT; // For fine tuning input voltage sense
+extern long voltMeterConstant; // For fine tuning input voltage sense
 
 #ifdef I2C_INPUT
 #define I2C_SLAVE_ADDRESS 24
